% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_pdf.R
\name{convert_pdf}
\alias{convert_pdf}
\title{Convert PDF to FST (line based)}
\usage{
convert_pdf(.dir)
}
\arguments{
\item{.dir}{The main directory for results}
}
\value{
FST files saved to to ".dir/fst"
}
\description{
Convert PDF to FST (line based)
}
