% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pdfs.R
\name{download_pdf}
\alias{download_pdf}
\title{Download PDF from Annualreport.com}
\usage{
download_pdf(.tab, .dir, .wait_time = 5, .mode = "wb")
}
\arguments{
\item{.tab}{A Dataframe created by get_links_map()}

\item{.dir}{The main directory for results}

\item{.wait_time}{Mean wait time between downloads (follows normal distribution with mean = .wait_time and sd = .2 * .wait_time)}

\item{.mode}{Download mode, default is "wb", see download.file() for more information}
}
\value{
\itemize{
\item PDFs will be downloaded to ".dir/pdf"\cr
\item Datframe saved as .xlsx that indicates success (".dir/results_download.xlsx)
}
}
\description{
Download PDF from Annualreport.com
}
