% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_lts.R
\name{extract_lts}
\alias{extract_lts}
\title{Extract Letter to shareholder}
\usage{
extract_lts(.tab, .dir, .default_length = 3)
}
\arguments{
\item{.tab}{A Dataframe with three columns: \cr
\itemize{
\item doc_id: Filename of the PDF\cr
\item start: The start page of the Letter section\cr
\item stop: The end page of the Letter section
}}

\item{.dir}{The main directory for results}

\item{.default_length}{If stop is empty, the number of pages to be extracted counting from start}
}
\value{
PDFs and text files saved to ".dir/lts"
}
\description{
Extract Letter to shareholder
}
