% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_lts.R
\name{find_lts_page_map}
\alias{find_lts_page_map}
\title{Find start page for Letter to Shareholders}
\usage{
find_lts_page_map(.dir, .phrases)
}
\arguments{
\item{.dir}{The main directory for results}

\item{.phrases}{A character vector with phrases}
}
\value{
A datframe with the page number of the start page of the Letter to shareholder section, saved to ".dir/letter_to_shareholder_pages.xlsx"
}
\description{
Find start page for Letter to Shareholders
}
