% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_links_map}
\alias{get_links_map}
\title{Get Links Report Links from AnnualReport.com}
\usage{
get_links_map(.tab, .col_url, .col_id, .col_ticker, .dir)
}
\arguments{
\item{.tab}{Dataframe with at least three columns \cr
\itemize{
\item URL column: Column with main URLs \cr
\item ID Column: Column with Unique Firm IDs \cr
\item Ricker Column: Column with unique Firm Ticker \cr
}}

\item{.col_url}{The URL column}

\item{.col_id}{The ID column}

\item{.col_ticker}{The Ticker column}

\item{.dir}{The main directory for results}
}
\value{
\itemize{
\item A dataframe with links to global environment and to ".dir/results_links.xlsx"\cr
\item A dataframe with potentiall errors to ".dir/errors_links.xlsx"
}
}
\description{
Get Links Report Links from AnnualReport.com
}
